backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pyjwt
requests
pytest
sqlalchemy
passlib[bcrypt]
python-dotenv
email-validator


